+++ go.mod
@@ -1,0 +1,18 @@
+module github.com/gravitational/shared-workflows
+
+go 1.22.4
+
+require (
+	github.com/alecthomas/kingpin/v2 v2.4.0
+	github.com/gravitational/trace v1.4.0
+	github.com/stretchr/testify v1.9.0
+)
+
+require (
+	github.com/alecthomas/units v0.0.0-20211218093645-b94a6e3cc137 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/xhit/go-str2duration/v2 v2.1.0 // indirect
+	golang.org/x/net v0.17.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
